make logic flow to util and move to file
